+++ app/views/homes/index.html.erb
@@ -1,0 +1,1 @@
+<%= javascript_pack_tag 'index' %>
